+++ core/js/src/test/scala/japgolly/clearconfig/JsTest.scala
@@ -1,0 +1,10 @@
+package japgolly.clearconfig
+
+import utest._
+
+object JsTest extends TestSuite {
+
+  override def tests = Tests {
+
+  }
+}