+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Replace Scalaz with Cats
 * Drop Scala 2.12 support
 * The JS module now includes ability to read environment (if in Node) and system properties
+* Added to `ConfigSource` instances:
+  * `expandInlineProperties(key)` for when a single value is the contents of a properties file
+  * `treatKeyDotsAsUnderscores`
+* `ConfigSource.environment` now longer calls `treatKeyDotsAsUnderscores` automatically.
+  Use `ConfigSource.environment[F].treatKeyDotsAsUnderscores` if required.
 
 ### 1.4.0